backend/Rentax/propiedades/serializers.py
from rest_framework import serializers
from .models import Propiedad

class PropiedadSerializer(serializers.ModelSerializer):
    class Meta:
        model = Propiedad
        fields = '__all__'  # incluye todos los campos del modelo